fix: Catch and swallow snackbar exception
For reasons not fully understood the root of a fragment's view might
not have relevant view from the activity set as its parent. This
causes `Snackbar.make()` to throw an exception, and crash.

See https://issuetracker.google.com/issues/228215869.

For now, "fix" this by swallowing the exception. Not showing the error
is better than crashing.