fix: Improve accessibility of ComposeActivity bottomsheets and buttons
The previous bottomsheets did set a minimum height for the menu items,
so they were less than the recommended 48dp minimum. Fix that to
improve the overall accessibility.

Always highlight the "visibilty" icon, to make it clear that it's
something that is set (even if to the default).

Show the visibility icon on the "Toot" button as an additional reminder
to the user.

Other changes:

- Use the "filled" style for all icons (the visibility icons had the
  "outlined" style)

- Use the `makeIcon` helper function.

- Use the `Status.Visibility` extension functions to determine the
  icon for each visibility type, reducing code duplication.