feat: Show preview image for video attachments
Video attachments may have a preview image. If they do, try and show
that image (from the cache) before the video loads, so the user has
something to look at. If there's no preview image, or it fails to load
fall back to a black screen with a progress spinner as normal.

This also provides a smooth transition when playing the first video
attachment, and fixes #598.