fix: Improve transitions in/out of playing video
Previous code didn't trigger the transition from `ViewMediaActivity`
when playing a video until the video had loaded. If the connection
was slow or had other issues this resulted in the video "sticking"
in the timeline until it loaded.

Change this, and trigger the transition immediately.

Fixes #598

While looking at this:

- Save the play/pause state of the video during a swipe, pause the
  video, and restore the state when the swipe is cancelled.
- Transition the entire video view, to improve the animated transition
  back to the activity.
- Remove the custom progress spinner, use the one provided by the
  player.
- Display the player controller via the layout XML instead of code