change: Adjust SwipeRefreshLayout offset in AccountActivity
The change in 5959d23d resulted in a stuck margin when scrolling up in
the account view.

Fix the `SwipeRefreshLayout` offset in a different way that doesn't
have that problem; move the spinner's start position to the top of the
screen (instead of the negative offset it normally has) and keep the
same relative offset from the end position.

Have it scale in on the drag to complete the effect.