fix: Only suggest upgrading to a version F-Droid has built
The F-Droid API can return a `suggestedVersionCode` that F-Droid has
not successfully built (e.g., there was a network issue preventing
F-Droid from fetching the source code at the time of the last build).

This meant users were being prompted to update when there was not a
built package to update to.

Fix this by verifying that `suggestedVersionCode` appears in the list of
packages. If it doesn't then fall back to the highest version code in
the list of packages.

Fixes #684