feat: Show autocomplete hashtags with usage counts, sort by popularity
When autocompleting hashtags while composing a status the previous
code showed the hashtags in the same order they're returned by the
server, with no additional information.

This doesn't allow the user to make an informed choice about which
hashtag might be better to use. For example, trying to choose between
"#nivenly" and "#NivenlyFoundation".

To fix that, include the hashtag's usage when receiving data from the
server. Sum that, and show it to the user in the hashtag list. Sort
the hashtags by popularity, most popular first.