fix: Render hashtags, mentions, and initial paras correctly in RTL
Previous code didn't set the textDirection for the status content,
so the first para of RTL text might be rendered incorrectly.

In addition, mentions and tags weren't BIDI wrapped, so would appear
as "foo@" and "foo#" in RTL statuses, instead of "@foo" and "#foo".

Fix both of these issues.

Fixes #870