fix: Tolerate missing fields on veganism.social
veganism.social is doing something weird that injects statuses into the
user's timeline that are missing the fields that are normally present
on authenticated statuses.

So far I've noticed:

- `reblogged`
- `favourited`
- `bookmarked`

on the top level `Status` are missing, and if the status contains a poll
the `voted` field is also missing.

Cover up this breakage by setting defaults for these fields.