fix: Restore video transitions on androidx.media3 1.4.0 and 1.4.1
androidx.media3 1.4.0-rc01 and above (at the time of writing) has a
bug that breaks shared element transitions with a `PlayerView` (see
https://github.com/androidx/media/issues/1594).

This can be worked around by setting `app:surface_type="texture_view"`.
This uses more power, but for the typical length of social media videos
this shouldn't be a problem at the moment.

Fixes #920.