Mostly convert AccountManager to be DB driven
Instead of storing the current state in memory, push changes down to
the database, and expose flows of those changes.

Convert consumers of AccountManager.activeAccount. The change so far
is bare minimum to get things building and working, some tests don't
work yet.

Filters and lists also don't work properly because of races with the
InstanceSwitchAuthInterceptor. That'll be fixed in the next commit,
where they'll be part of the Account info and fetched when the account
info is fetched during the account switch process.