feat: Add option to save attachments to per-account folders
The existing code downloaded any attachments to the user's "Downloads"
folder. If the user is logged in with several accounts these downloads
will be mixed up together.

Fix this by adding a new preference that allows the user to specify
the downloads should be placed in a sub-folder per account, named
after the account.

To do this:

- Add an interface for enums that can be used as preferences, with
  properties for the string resource to display and the value to store.
- Add `EnumListPreference`, a `ListPreference` that allows the user to
  choose between different enum values.
- Add a `DownloadLocation` enum and preference key so the user can
  choose the location.
- Add a `core.domain` module, with a use case for downloading URLs
  that respect's the user's download preference. Use this use-case
  everywhere that files are currently downloaded.

Fixes #938