feat: Add option to download media to per-sender directories
Update DownloadUrlUseCase with a parameter to specify the account that
"owns" the media. This is either the account that posted the status,
or the account being viewed (e.g., if downloading an account's header
image).

Add a new `DownloadLocation` enum constant to download to directories
named after that account.

Pass this information through at the call sites.

Fixes #938