test: Ensure that all relevant MastodonAPI methods are stubbed
Tests that rely on stubbing the MastodonAPI may have incorrect
behaviour if some of the API methods called by the system-under-test
are not stubbed. This can be difficult to track down.

Make this easier by adding a default answer for the MastodonAPI used
in tests. Any un-stubbed methods will throw an `AssertionError` with
enough information to identify the method that should be stubbed and
where it was called from.