feat: Add "Don't ask again" option when confirming the posting language
For some users this feature has too many false positives, particularly
when emojis are used.

Provide a preference that enables/disables the feature, default to
"enable".

When the user is prompted the dialog has a third "Don't ask again"
option. If chosen the preference is set to prevent future prompts
and the status is sent as-is.

Fixes #893