feat: Add accessibility options for tab justification and content
Provide two new lab preferences for controlling the layout and content
of main navigation tabs.

Tabs can now be justfied to start, end, or fully (if room). start/end
justification may put the tabs closer to the user's fingers, depending
on how they hold the device. Fully justified uses the full width of
the tab bar (if the tabs don't require scrolling).

The content can be set to one of:

- Icon only (previous behaviour)
- Text only
- Icon with text beside
- Icon with text below