feat: Show text counter on caption dialog, request focus
Previous layout didn't enable the character counter for entered text,
so it was difficult for the user to see how much they had used.

Fix that. To get the layout to play nicely switch to a ConstraintLayout
with a chain to split the dialog roughly 50/50 between the image and
the caption views.

While I'm here:

- Remove the margins around the image, so it's clearer, and is more
consistent with the "set focus" dialog.
- Set focus on the description immediately, so the keyboard pops up.
- Remove focusability from the image, since the user shouldn't be
able to interact with it.