feat: Show share sheet when long clicking links/hashtags/mentions
This makes it easier to share links from a regular timeline (where
text is not selectable).

Listen for long clicks on spans by launching a delayed runnable that
waits `getLongPressTimeout()` ms before sending an intent that shows
the share sheet for the clicked span.

If the span's link is not the same as the text (e.g., it's a mention
or a hashtag) then the text of the span is also sent as the title to
show in the share sheet.

While I'm here, the official LinkMovementMethodCompat does the same job
as NoTrailingSpaceLinkMovementMethod, so delete the latter code in
favour of the former.

Fixes #695